Move toc Token to token.ts
