+++ packages/toc/src/generators/markdown/get_headings.ts
@@ -2,7 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 
 import { generateNumbering } from '../../utils/generate_numbering';
-import { INumberedHeading } from '../../utils/headings';
+import { INumberedHeading } from '../../tokens';
 import { INumberingDictionary } from '../../utils/numbering_dictionary';
 import { parseHeading } from '../../utils/parse_heading';
