+++ packages/toc/src/generators/markdown/index.ts
@@ -8,17 +8,16 @@ import {
   IMarkdownViewerTracker,
   MarkdownDocument
 } from '@jupyterlab/markdownviewer';
+import { ISettingRegistry } from '@jupyterlab/settingregistry';
 import { ITranslator, nullTranslator } from '@jupyterlab/translation';
-import { TableOfContentsRegistry as Registry } from '../../registry';
 import { TableOfContents } from '../../toc';
-import { INumberedHeading } from '../../utils/headings';
+import { INumberedHeading } from '../../tokens';
 import { isMarkdown } from '../../utils/is_markdown';
+import { getHeadings } from './get_headings';
+import { getRenderedHeadings } from './get_rendered_headings';
 import { OptionsManager } from './options_manager';
 import { render } from './render';
 import { toolbar } from './toolbar_generator';
-import { getHeadings } from './get_headings';
-import { getRenderedHeadings } from './get_rendered_headings';
-import { ISettingRegistry } from '@jupyterlab/settingregistry';
 
 /**
  * Returns a boolean indicating whether this ToC generator is enabled.
@@ -222,3 +221,4 @@ function createRenderedMarkdownGenerator(
  */
 export { createMarkdownGenerator };
 export { createRenderedMarkdownGenerator };
+
