+++ packages/toc/src/generators/markdown/options_manager.ts
@@ -3,8 +3,8 @@
 
 import { ISanitizer } from '@jupyterlab/apputils';
 import { ITranslator, nullTranslator } from '@jupyterlab/translation';
-import { TableOfContentsRegistry as Registry } from '../../registry';
 import { TableOfContents } from '../../toc';
+import { TableOfContentsRegistry as Registry } from '../../tokens';
 
 /**
  * Interface describing constructor options.
