+++ packages/toc/src/generators/markdown/render.tsx
@@ -2,7 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 
 import * as React from 'react';
-import { INumberedHeading } from '../../utils/headings';
+import { INumberedHeading } from '../../tokens';
 import { sanitizerOptions } from '../../utils/sanitizer_options';
 import { OptionsManager } from './options_manager';
