+++ packages/toc/src/generators/notebook/get_markdown_heading.ts
@@ -3,7 +3,7 @@
 
 import { Cell } from '@jupyterlab/cells';
 import { generateNumbering } from '../../utils/generate_numbering';
-import { INotebookHeading, RunningStatus } from '../../utils/headings';
+import { INotebookHeading, RunningStatus } from '../../tokens';
 import { INumberingDictionary } from '../../utils/numbering_dictionary';
 import { parseHeading } from '../../utils/parse_heading';
