+++ packages/toc/src/generators/notebook/is_heading_filtered.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { INotebookHeading } from '../../utils/headings';
+import { INotebookHeading } from '../../tokens';
 
 /**
  * Returns a boolean indicating whether a heading is filtered out by selected tags.
