+++ packages/toc/src/generators/python/index.ts
@@ -3,8 +3,7 @@
 
 import { IDocumentWidget } from '@jupyterlab/docregistry';
 import { FileEditor, IEditorTracker } from '@jupyterlab/fileeditor';
-import { TableOfContentsRegistry as Registry } from '../../registry';
-import { IHeading } from '../../utils/headings';
+import { IHeading, TableOfContentsRegistry as Registry } from '../../tokens';
 import { render } from './render';
 
 /**
