+++ packages/toc/src/registry.ts
@@ -1,30 +1,17 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { IWidgetTracker } from '@jupyterlab/apputils';
-import { Token } from '@lumino/coreutils';
 import { ISignal, Signal } from '@lumino/signaling';
 import { Widget } from '@lumino/widgets';
-import { IHeading, INotebookHeading } from './utils/headings';
-
-/**
- * Interface describing the table of contents registry.
- *
- * @private
- */
-export interface ITableOfContentsRegistry extends TableOfContentsRegistry {}
-
-/**
- * Table of contents registry token.
- */
-export const ITableOfContentsRegistry = new Token<TableOfContentsRegistry>(
-  '@jupyterlab/toc:ITableOfContentsRegistry'
-);
+import {
+  ITableOfContentsRegistry,
+  TableOfContentsRegistry as ToCNamespace
+} from './tokens';
 
 /**
  * Class for registering widgets for which we can generate a table of contents.
  */
-export class TableOfContentsRegistry {
+export class TableOfContentsRegistry implements ITableOfContentsRegistry {
   /**
    * Finds a table of contents generator for a widget.
    *
@@ -35,7 +22,7 @@ export class TableOfContentsRegistry {
    * @param widget - widget
    * @returns table of contents generator
    */
-  find(widget: Widget): TableOfContentsRegistry.IGenerator | undefined {
+  find(widget: Widget): ToCNamespace.IGenerator | undefined {
     for (let i = 0; i < this._generators.length; i++) {
       const gen = this._generators[i];
       if (gen.tracker.has(widget)) {
@@ -52,13 +39,13 @@ export class TableOfContentsRegistry {
    *
    * @param generator - table of contents generator
    */
-  add(generator: TableOfContentsRegistry.IGenerator): void {
+  add(generator: ToCNamespace.IGenerator): void {
     if (generator.collapseChanged) {
       // If there is a collapseChanged for a given generator, propagate the arguments through the registry's signal
       generator.collapseChanged.connect(
         (
-          sender: TableOfContentsRegistry.IGenerator<Widget>,
-          args: TableOfContentsRegistry.ICollapseChangedArgs
+          sender: ToCNamespace.IGenerator<Widget>,
+          args: ToCNamespace.ICollapseChangedArgs
         ) => {
           this._collapseChanged.emit(args);
         }
@@ -67,124 +54,13 @@ export class TableOfContentsRegistry {
     this._generators.push(generator);
   }
 
-  get collapseChanged(): ISignal<
-    this,
-    TableOfContentsRegistry.ICollapseChangedArgs
-  > {
+  get collapseChanged(): ISignal<this, ToCNamespace.ICollapseChangedArgs> {
     return this._collapseChanged;
   }
 
   private _collapseChanged: Signal<
     this,
-    TableOfContentsRegistry.ICollapseChangedArgs
-  > = new Signal<this, TableOfContentsRegistry.ICollapseChangedArgs>(this);
-  private _generators: TableOfContentsRegistry.IGenerator[] = [];
-}
-
-/**
- * Static registry methods.
- */
-export namespace TableOfContentsRegistry {
-  /**
-   * Abstract class for managing options affecting how a table of contents is generated for a particular widget type.
-   */
-  export interface IOptionsManager {}
-
-  /**
-   * Interface for the arguments needed in the collapse signal of a generator
-   */
-  export interface ICollapseChangedArgs {
-    /**
-     * Boolean indicating whether the given heading is collapsed in ToC
-     */
-    collapsedState: boolean;
-
-    /**
-     * Heading that was involved in the collapse event
-     */
-    heading: IHeading;
-
-    /**
-     * Type of file that the given heading was produced from
-     */
-    tocType: string;
-  }
-
-  /**
-   * Interface describing a widget table of contents generator.
-   */
-  export interface IGenerator<W extends Widget = Widget> {
-    /**
-     * Widget instance tracker.
-     */
-    tracker: IWidgetTracker<W>;
-
-    /**
-     * Returns a boolean indicating whether we can generate a ToC for a widget.
-     *
-     * ## Notes
-     *
-     * -   By default, we assume ToC generation is enabled if the widget is hosted in `tracker`.
-     * -   However, a user may want to add additional checks (e.g., only generate a ToC for text files only if they have a given MIME type).
-     *
-     * @param widget - widget
-     * @returns boolean indicating whether we can generate a ToC for a widget
-     */
-    isEnabled?: (widget: W) => boolean;
-
-    /**
-     * Boolean indicating whether a document uses LaTeX typesetting.
-     *
-     * @default false
-     */
-    usesLatex?: boolean;
-
-    /**
-     * Options manager.
-     *
-     * @default undefined
-     */
-    options?: IOptionsManager;
-
-    /**
-     * Signal to indicate that a collapse event happened to this heading
-     * within the ToC.
-     */
-    collapseChanged?: ISignal<IOptionsManager, ICollapseChangedArgs>;
-
-    /**
-     * Returns a JSX element for each heading.
-     *
-     * ## Notes
-     *
-     * -   If not present, a default renderer will be used.
-     *
-     * @param item - heading
-     * @param toc - list of headings
-     * @returns JSX element
-     */
-    itemRenderer?: (
-      item: IHeading,
-      toc: INotebookHeading[]
-    ) => JSX.Element | null;
-
-    /**
-     * Returns a toolbar component.
-     *
-     * ## Notes
-     *
-     * -   If not present, no toolbar is generated.
-     *
-     * @returns toolbar component
-     */
-    toolbarGenerator?: () => any;
-
-    /**
-     * Returns a list of headings.
-     *
-     * @param widget - widget
-     * @returns list of headings
-     */
-    generate(widget: W, options?: IOptionsManager): IHeading[];
-  }
+    ToCNamespace.ICollapseChangedArgs
+  > = new Signal<this, ToCNamespace.ICollapseChangedArgs>(this);
+  private _generators: ToCNamespace.IGenerator[] = [];
 }
